b2b_e_commerce/requirements.txt
asgiref==3.7.1
attrs==23.1.0
Django==4.2.1
django-autoslug==1.9.9
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
inflection==0.5.1
jsonschema==4.17.3
phonenumbers==8.13.11
Pillow==9.5.0
psycopg2==2.9.6
PyJWT==2.7.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
